Reduce CheckDecode memory allocs.
Profiling has shown the nearly 6.5% of the total allocations in dcrd are
the result of CheckDecode.

Consequently, this updates the CheckDecode function to make use of the
new zero alloc blake256.Sum256 function to reduce the memory allocations
needed to perform the decode.  It also improves the readability of the
function a bit while here.

The following is a before and after comparison of a typical decode.

benchmark             old ns/op    new ns/op    delta
------------------------------------------------------
BenchmarkCheckDecode  4176         4088         -2.11%

benchmark             old allocs   new allocs   delta
-------------------------------------------------------
BenchmarkCheckDecode   12          9            -25.00%

benchmark             old bytes    new bytes    delta
-------------------------------------------------------
BenchmarkCheckDecode  472          264          -44.07%