Optimize Decode.
Profiling has shown the nearly 6.5% of the total allocations in dcrd are
the result of CheckDecode, most of which are via Decode.

Consequently, this updates the Decode function to reduce the memory
allocations needed to perform the decode.  It also improves the
readability of the function a bit while here.

The following is a before and after comparison of a typical decode
including the results of CheckDecode since it uses Decode.

benchmark               old ns/op    new ns/op    delta
----------------------------------------------------------
BenchmarkBase58Decode   4075         1032         -74.67%
BenchmarkCheckDecode    4104         1399         -65.91%

benchmark               old allocs   new allocs   delta
----------------------------------------------------------
BenchmarkBase58Decode   8            1            -87.50%
BenchmarkCheckDecode    9            2            -77.78%

benchmark               old bytes    new bytes    delta
----------------------------------------------------------
BenchmarkBase58Decode   232          48           -79.31%
BenchmarkCheckDecode    264          80           -69.70%